Smoothed out some communication between front and back

+++ ClientApp/src/API.js
@@ -5,30 +5,40 @@ export default {
         // Get
         getAssets: () =>  axios.get('api/asset'),
         getAsset: query => {
-            console.log("query: ", query);
+            console.log("getAsset query: ", query);
+            if (!query) return this.messages.badInput;
             return axios.get(`api/asset/${query}`);
         },
         getShares: query => {
-            console.log("query: ", query);
-            return axios.get(`api/asset/shares/${query}`);
+            console.log("getShares query: ", query ? query : "cash");
+            return axios.get(query ? `api/asset/shares/${query}` : 'api/asset/shares');
+        },
+        getPlayerShares: query => {
+            console.log("getPlayerShares query: ", query);
+            if (!query) return this.messages.badInput;
+            return axios.get(`api/asset/player-shares/${query.id}/${query.type}`);
         },
         // Post
         createAsset: query => {
-            console.log("query: ", query);
+            console.log("createAsset query: ", query);
+            if (!query) return this.messages.badInput;
             return axios.post(`api/asset`, query);
         },
         addShares: query => {
-            console.log("query: ", query);
+            console.log("addShares query: ", query);
+            if (!query) return this.messages.badInput;
             return axios.post(`api/asset/${query.id}/${query.qty}`)
         },
         // Put
         updateAsset: query => {
-            console.log("query: ", query);
+            console.log("updateAsset query: ", query);
+            if (!query) return this.messages.badInput;
             return axios.put(`api/asset/${query.id}`, query.changeSet);
         },
         // Delete
         deleteAsset: query => {
-            console.log("query: ", query);
+            console.log("deleteAsset query: ", query);
+            if (!query) return this.messages.badInput;
             return axios.delete(`api/asset/${query}`);
         }
     },
@@ -36,41 +46,55 @@ export default {
         // Get
         getPlayers: () => axios.get('api/player'),
         getPlayer: query => {
-            console.log("query: ", query);
+            console.log("getPlayer query: ", query);
+            if (!query) return this.messages.badInput;
             return axios.get(`api/player/${query}`);
         },
         // Post
         createPlayer: query => {
-            console.log("query: ", query);
+            console.log("createPlayer query: ", query);
+            if (!query) return this.messages.badInput;
             return axios.post('api/player/', query);
         },
         addSharesToPlayer: query => {
             console.log("add shares query: ", query);
+            if (!query) return this.messages.badInput;
             return axios.post(`api/player/${query.playerId}/${query.assetId}/${query.qty}`, query.changeSet);
         },
         // Put
         updatePlayer: query => {
-            console.log("query: ", query);
+            console.log("updatePlayer query: ", query);
+            if (!query) return this.messages.badInput;
             return axios.put(`api/player/${query.id}`, query.changeSet);
         },
         clearPortfolio: query => {
-            console.log("query: ", query);
+            console.log("clearPortfolio query: ", query);
+            if (!query) return this.messages.badInput;
             return axios.put(`api/player/${query}`);
         },
         clearWallet: query => {
-            console.log("query: ", query);
+            console.log("clearWallet query: ", query);
+            if (!query) return this.messages.badInput;
             return axios.put(`api/player/${query}`);
         },
         // Delete
         deletePlayer: query => {
-            console.log("query: ", query);
+            console.log("deletePlayer query: ", query);
+            if (!query) return this.messages.badInput;
             return axios.delete(`api/player/${query}`);
         }
     },
     transactions: {
         submitTrade: query => {
-            console.log("query: ", query);
+            console.log("submit trade query: ", query);
+            if (!query) return this.messages.badInput;
             return axios.post('api/transaction', query);
         }
+    },
+    initialize: () => {
+        axios.get('/game/start/initialize');
+    },
+    messages: {
+        badInput: "Huh? Didn't hear that"
     }
 };
